Se realiza cambios en la ubicacion de los archivos y se elimina las tabs de centro gestor
