+++ src/features/functionality/interfaces/Budgets.ts
@@ -15,11 +15,3 @@ export interface IBudgets {
   entity?: IEntities
 }
 
-export interface IFilterBudgets {
-  page: number;
-  perPage: number;
-  entity?: number;
-  ejercise?: number;
-  number:number;
-  denomination:string;
-}
