+++ src/features/managementCenter/interfaces/FilterBudgets.ts
@@ -1,0 +1,9 @@
+export interface IFilterBudgets {
+    page: number;
+    perPage: number;
+    entity?: number;
+    ejercise?: number;
+    number:number;
+    denomination:string;
+}
+
