+++ src/features/managementCenter/management-center-routes.tsx
@@ -5,8 +5,8 @@ import PrivateRoute from '../../../../sapiencia-core.front/src/common/components
 
 export const ManagementCenterRoutes = () => {
 
-    const ManagementCenterAdditionalPage =  lazy(() => import("../functionality/pages/management-center-additional.page"));
-    const ManagementCenterPage = lazy(() => import("../functionality/pages/management-center.page"));
+    const ManagementCenterAdditionalPage =  lazy(() => import("./pages/management-center-additional.page"));
+    const ManagementCenterTransferPage = lazy(() => import("./pages/management-center-transfer.page"));
 
   return (
     <Routes>
@@ -16,11 +16,7 @@ export const ManagementCenterRoutes = () => {
         />
         <Route
             path={"/traslado"}
-            element={<ManagementCenterPage/> }
-        />
-         <Route
-            path={"/traslado/:option"}
-            element={<ManagementCenterPage/> }
+            element={<ManagementCenterTransferPage/> }
         />
         <Route
             path={"/disminucion"}
